chore(JsonConverter): refactor JsonConverter into JsonConverterSerializer & JsonConverterDeserializer
