+++ src/mapping/JsonValidators.ts
@@ -1,0 +1,7 @@
+export type JsonValidator = (value: any) => boolean;
+
+export class JsonValidators {
+
+    public static required: JsonValidator = (value: any) => value !== null;
+
+}
