+++ src/mapping/PropertyMapping.ts
@@ -1,5 +1,8 @@
+import {JsonValidator} from "../decorators/jsonProperty";
+
 export interface PropertyMapping {
     name: string;
     serializedName: string;
     type?: any;
+    validators: JsonValidator[];
 }
