+++ src/mapping/api.ts
@@ -1,3 +1,4 @@
 export * from './PropertyMapping';
+export * from './JsonValidators';
 export * from './TypeMapping';
 export * from './JsonConverterMapper';
