+++ src/samples/models/Vehicle.ts
@@ -4,17 +4,18 @@ import {Passenger} from "./Passenger";
 import {Color} from "./Color";
 import {VehicleOptions} from "./VehicleOptions";
 import {Enum, EnumStrategy} from "../../type/Enum";
+import {JsonValidators} from "../../mapping/JsonValidators";
 
 @jsonObject({discriminatorProperty: 'type'})
 export abstract class Vehicle {
 
-    @jsonProperty('type', String)
+    @jsonProperty('type', String, [JsonValidators.required])
     public _type: string;
 
-    @jsonProperty('id', Number)
+    @jsonProperty('id', Number, [JsonValidators.required])
     public _id: number;
 
-    @jsonProperty('name', String)
+    @jsonProperty('name', String, [JsonValidators.required])
     public _name: string;
 
     @jsonProperty('passengers', [Passenger])
